Another attempt to fix viaversion cancelling packets.
User#closeInventory added

+++ api/src/main/java/com/github/retrooper/packetevents/wrapper/play/server/WrapperPlayServerCloseWindow.java
@@ -11,13 +11,13 @@ public class WrapperPlayServerCloseWindow extends PacketWrapper<WrapperPlayServe
         super(event);
     }
 
-    public WrapperPlayServerCloseWindow(byte id) {
+    public WrapperPlayServerCloseWindow(int id) {
         super(PacketType.Play.Server.CLOSE_WINDOW);
         this.windowId = id;
     }
 
     public WrapperPlayServerCloseWindow() {
-        this((byte)0);
+        this(0);
     }
 
     @Override
